ch03-data-pipeline/requirements.txt
python-docx>=1.1.0
openpyxl>=3.1.2
xlrd>=2.0.1
pdfplumber>=0.10.3
